% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_challenges.R
\name{cr_get_challenges}
\alias{cr_get_challenges}
\title{Get Challenges}
\usage{
cr_get_challenges(
  limit = NULL,
  after = NULL,
  before = NULL,
  key = cr_get_key()
)
}
\arguments{
\item{limit}{number of items returned in the response.}

\item{after}{return only items that occur after this marker.}

\item{before}{return only items that occur before this marker.}

\item{key}{Required. API key. See \url{https://developer.clashroyale.com/#/documentation}
Default: \code{cr_get_key()}}
}
\value{
tibble of card info
}
\description{
Get Challenges
}
\examples{
\dontshow{if (royale::cr_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_get_challenges()
\dontshow{\}) # examplesIf}
}
\concept{challenges}
