% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clan.R
\name{cr_get_clan}
\alias{cr_get_clan}
\title{Get a clan}
\usage{
cr_get_clan(clan = "99R2PQVR", key = cr_get_key())
}
\arguments{
\item{clan}{Required. Clan tag. Default: 99R2PQVR}

\item{key}{Required. API key. See \url{https://developer.clashroyale.com/#/documentation}
Default: cr_get_key}
}
\value{
a \code{tibble} with each row as a player
}
\description{
Gets full clan details
}
\examples{
\dontshow{if (royale::cr_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_get_clan('99R2PQVR')
\dontshow{\}) # examplesIf}
}
\concept{clan}
