% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clan_members.R
\name{cr_get_clan_members}
\alias{cr_get_clan_members}
\title{Get current clan members}
\usage{
cr_get_clan_members(
  clan = "99R2PQVR",
  limit = NULL,
  after = NULL,
  before = NULL,
  key = cr_get_key()
)
}
\arguments{
\item{clan}{Required. Clan tag.
Default: 99R2PQVR}

\item{limit}{number of items returned in the response.}

\item{after}{return only items that occur after this marker.}

\item{before}{return only items that occur before this marker.}

\item{key}{Required. Clash Royale API key. See <https://developer.clashroyale.com/#/documentation>
Default: cr_get_key}
}
\value{
`tibble` with list columns for clan, clans, and history.
}
\description{
Get current clan members
}
\examples{
\dontshow{if (royale::cr_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_get_clan_members('99R2PQVR')
\dontshow{\}) # examplesIf}
}
\concept{clan}
