% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clan_search.R
\name{cr_get_clan_search}
\alias{cr_get_clan_search}
\title{Search Clans}
\usage{
cr_get_clan_search(
  name = NULL,
  location_id = NULL,
  min_members = NULL,
  max_members = NULL,
  min_score = NULL,
  limit = NULL,
  after = NULL,
  before = NULL,
  key = cr_get_key()
)
}
\arguments{
\item{name}{Search clans by name}

\item{location_id}{Filter by clan location identifier}

\item{min_members}{Filter by minimum number of clan members}

\item{max_members}{Filter by maximum number of clan members}

\item{min_score}{Filter by minimum amount of clan score}

\item{limit}{number of items returned in the response.}

\item{after}{return only items that occur after this marker.}

\item{before}{return only items that occur before this marker.}

\item{key}{Required. API key. See \url{https://developer.clashroyale.com/#/documentation}
Default: cr_get_key}
}
\value{
a \code{tibble} with each row as a player
}
\description{
Search Clans
}
\examples{
\dontshow{if (royale::cr_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_get_clan_search('Legend Dynasty!')
\dontshow{\}) # examplesIf}
}
\concept{clan}
