% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clan_war_log.R
\name{cr_get_clan_war_log}
\alias{cr_get_clan_war_log}
\title{Gets Clan War Log}
\usage{
cr_get_clan_war_log(clan = "99R2PQVR", key = cr_get_key())
}
\arguments{
\item{clan}{Required. Clan tag.
Default: 99R2PQVR}

\item{key}{Required. Clash Royale API key. See https://developer.clashroyale.com/#/documentation
Default: cr_get_key}
}
\value{
Returns clan tibble
}
\description{
Gets Clan War Log
}
