% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_war.R
\name{cr_get_current_war}
\alias{cr_get_current_war}
\title{Gets current clan war}
\usage{
cr_get_current_war(clan = "99R2PQVR", key = cr_get_key())
}
\arguments{
\item{clan}{Required. Clan tag.
Default: 99R2PQVR}

\item{key}{Required. Clash Royale API key. See https://developer.clashroyale.com/#/documentation
Default: cr_get_key}
}
\value{
Returns clan war log
}
\description{
Gets current clan war
}
