% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_ranking_players.R
\name{cr_get_global_ranking_players}
\alias{cr_get_global_ranking_players}
\title{Get Global Player Rankings for a Season}
\usage{
cr_get_global_ranking_players(
  season,
  limit = NULL,
  after = NULL,
  before = NULL,
  key = cr_get_key()
)
}
\arguments{
\item{season}{Required. Season ID, such as '2022-09'}

\item{limit}{number of items returned in the response.}

\item{after}{return only items that occur after this marker.}

\item{before}{return only items that occur before this marker.}

\item{key}{Required. API key. See \url{https://developer.clashroyale.com/#/documentation}
Default: \code{cr_get_key()}}
}
\value{
tibble of card info
}
\description{
Get Global Player Rankings for a Season
}
\examples{
\dontshow{if (royale::cr_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_get_global_ranking_players('2022-09', limit = 10)
\dontshow{\}) # examplesIf}
}
\concept{locations}
