% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_ranking_tournament.R
\name{cr_get_global_ranking_tournament}
\alias{cr_get_global_ranking_tournament}
\title{Get Global Tournament Rankings}
\usage{
cr_get_global_ranking_tournament(
  tournament,
  limit = NULL,
  after = NULL,
  before = NULL,
  key = cr_get_key()
)
}
\arguments{
\item{tournament}{Tournament tag. Required.}

\item{limit}{number of items returned in the response.}

\item{after}{return only items that occur after this marker.}

\item{before}{return only items that occur before this marker.}

\item{key}{Required. API key. See \url{https://developer.clashroyale.com/#/documentation}
Default: \code{cr_get_key()}}
}
\value{
tibble of card info
}
\description{
Get Global Tournament Rankings
}
\examples{
\dontshow{if (royale::cr_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# needs an ongoing tournament at time of request
try({cr_get_global_ranking_tournament('URP8YYUU0')})
\dontshow{\}) # examplesIf}
}
\concept{locations}
