% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_ranking_clans.R
\name{cr_get_local_ranking_clans}
\alias{cr_get_local_ranking_clans}
\title{Get Clan Rankings for a Location}
\usage{
cr_get_local_ranking_clans(
  location,
  limit = NULL,
  after = NULL,
  before = NULL,
  key = cr_get_key()
)
}
\arguments{
\item{location}{Required. Location ID from \code{cr_get_locations()}.}

\item{limit}{number of items returned in the response.}

\item{after}{return only items that occur after this marker.}

\item{before}{return only items that occur before this marker.}

\item{key}{Required. API key. See \url{https://developer.clashroyale.com/#/documentation}
Default: \code{cr_get_key()}}
}
\value{
tibble of card info
}
\description{
Get Clan Rankings for a Location
}
\examples{
\dontshow{if (royale::cr_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_get_local_ranking_clans(57000001, limit = 10)
\dontshow{\}) # examplesIf}
}
\concept{locations}
