% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location.R
\name{cr_get_location}
\alias{cr_get_location}
\title{Get One Location}
\usage{
cr_get_location(location, key = cr_get_key())
}
\arguments{
\item{location}{Required. Location ID from \code{cr_get_locations()}.}

\item{key}{Required. API key. See \url{https://developer.clashroyale.com/#/documentation}
Default: \code{cr_get_key()}}
}
\value{
tibble of card info
}
\description{
Get One Location
}
\examples{
\dontshow{if (royale::cr_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_get_location(57000001)
\dontshow{\}) # examplesIf}
}
\concept{locations}
