% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player.R
\name{cr_get_player}
\alias{cr_get_player}
\title{Get a single player}
\usage{
cr_get_player(tag = "JYJQC88", key = cr_get_key())
}
\arguments{
\item{tag}{Required. The player to return.
Default: JYJQC88}

\item{key}{Required. API key. See \url{https://developer.clashroyale.com/#/documentation}
Default: \code{cr_get_key()}}
}
\value{
Returns a player's info as a tibble
}
\description{
Gets the player details for one player.
}
\examples{
#TODO
}
