% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seasons.R
\name{cr_get_seasons}
\alias{cr_get_seasons}
\title{Get Global Seasons}
\usage{
cr_get_seasons(key = cr_get_key())
}
\arguments{
\item{key}{Required. API key. See \url{https://developer.clashroyale.com/#/documentation}
Default: \code{cr_get_key()}}
}
\value{
a \code{tibble} of season info
}
\description{
This returns the current season if it's a valid season.
}
\examples{
\dontshow{if (royale::cr_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_get_seasons()
\dontshow{\}) # examplesIf}
}
\concept{locations}
