% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tournament.R
\name{cr_get_tournament}
\alias{cr_get_tournament}
\title{Get Tournament}
\usage{
cr_get_tournament(tournament, key = cr_get_key())
}
\arguments{
\item{tournament}{Tournament tag. Required.}

\item{key}{Required. API key. See \url{https://developer.clashroyale.com/#/documentation}
Default: cr_get_key}
}
\value{
a \code{tibble} with each row as a player
}
\description{
Get Tournament
}
\examples{
\dontshow{if (royale::cr_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tournament <- cr_get_tournament_search('a')$tag[1]
cr_get_tournament(tournament)
\dontshow{\}) # examplesIf}
}
\concept{tournament}
