% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tournament_search.R
\name{cr_get_tournament_search}
\alias{cr_get_tournament_search}
\title{Search Tournaments}
\usage{
cr_get_tournament_search(
  name = NULL,
  limit = NULL,
  after = NULL,
  before = NULL,
  key = cr_get_key()
)
}
\arguments{
\item{name}{Search clans by name}

\item{limit}{number of items returned in the response.}

\item{after}{return only items that occur after this marker.}

\item{before}{return only items that occur before this marker.}

\item{key}{Required. API key. See \url{https://developer.clashroyale.com/#/documentation}
Default: cr_get_key}
}
\value{
a \code{tibble} with each row as a player
}
\description{
Search Tournaments
}
\examples{
\dontshow{if (royale::cr_has_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cr_get_tournament_search('clash')
\dontshow{\}) # examplesIf}
}
\concept{tournament}
