% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version.R
\name{cr_get_version}
\alias{cr_get_version}
\title{Get RoyaleAPI version}
\usage{
cr_get_version(key = cr_get_key())
}
\arguments{
\item{key}{Required. API key. See https://developer.clashroyale.com/#/documentation
Default: cr_get_key}
}
\value{
Returns RoyaleAPI version
}
\description{
Gets the current version for RoyaleAPI
}
