% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{cr_set_key}
\alias{cr_set_key}
\title{Set Clash Royale API key}
\usage{
cr_set_key(key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{key}{Required. Your Clash Royale API key. See <https://developer.clashroyale.com/#/getting-started>.}

\item{overwrite}{Defaults to FALSE. Boolean. Should existing `CR_API_KEY` in Renviron be overwritten?}

\item{install}{Defaults to FALSE. Boolean. Should this be added '~/.Renviron' file?}
}
\value{
None
}
\description{
Sets your Clash Royale API key for default use.
}
\examples{
\dontrun{
cr_set_key('1234')
}

}
