% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clan_war_log.R
\name{get_clan_war_log}
\alias{get_clan_war_log}
\title{Gets clan war log}
\usage{
get_clan_war_log(clan = "99R2PQVR", key = get_Royale())
}
\arguments{
\item{clan}{Required. Clan tag.
Default: 99R2PQVR}

\item{key}{Required. Clash Royale API key. See \href{docs.royaleapi.com/#/}{RoyaleAPI.} \cr
Default: get_Royale}
}
\value{
Returns clan war log
}
\description{
Gets clan war log
}
