% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_riverrace.R
\name{get_current_riverrace}
\alias{get_current_riverrace}
\title{Gets current River Race}
\usage{
get_current_riverrace(clan = "99R2PQVR", key = cr_get_key())
}
\arguments{
\item{clan}{Required. Clan tag.
Default: 99R2PQVR}

\item{key}{Required. Clash Royale API key. See https://developer.clashroyale.com/#/documentation
Default: cr_get_key}
}
\value{
Returns river race as ___
}
\description{
Gets current River Race
}
