% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_war.R
\name{get_current_war}
\alias{get_current_war}
\title{Gets current clan war}
\usage{
get_current_war(clan = "99R2PQVR", key = get_Royale())
}
\arguments{
\item{clan}{Required. Clan tag.
Default: 99R2PQVR}

\item{key}{Required. Clash Royale API key. See \href{docs.royaleapi.com/#/}{RoyaleAPI.} \cr
Default: get_Royale}
}
\value{
Returns clan war log
}
\description{
Gets current clan war
}
