% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoints.R
\name{get_endpoints}
\alias{get_endpoints}
\title{Get Clash Royale API endpoints}
\usage{
get_endpoints(key = get_Royale())
}
\arguments{
\item{key}{Required. API key. See \href{https://developer.clashroyale.com/#/}{Clash Royale Api} \cr
Default: get_Royale}
}
\value{
Returns available endpoints
}
\description{
Gets the available endpoints for Clash Royale APi
}
