% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player.R
\name{get_player}
\alias{get_player}
\title{Get a single player}
\usage{
get_player(tag = "JYJQC88", key = get_Royale())
}
\arguments{
\item{tag}{Required. The player to return.
Default: JYJQC88}

\item{key}{Required. API key. See \href{docs.royaleapi.com/#/}{RoyaleAPI.} \cr
Default: get_Royale}
}
\value{
Returns a player's info as a tibble
}
\description{
Gets the player details for one player.
}
