% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_version.R
\name{get_version}
\alias{get_version}
\title{Get RoyaleAPI version}
\usage{
get_version(key = get_Royale())
}
\arguments{
\item{key}{Required. API key. See \href{docs.royaleapi.com/#/}{RoyaleAPI.} \cr
Default: get_Royale}
}
\value{
Returns RoyaleAPI version
}
\description{
Gets the current version for RoyaleAPI
}
