% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{key}
\alias{key}
\alias{cr_get_key}
\alias{cr_has_key}
\title{Get Clash Royale API key}
\usage{
cr_get_key()

cr_has_key()
}
\value{
string key, invisibly otherwise logical
}
\description{
Retrieves Clash Royale API key set by \code{cr_set_key()} if \code{cr_get_key()} or indicates if
one is set if \code{cr_has_key()}.
}
\examples{
cr_has_key()
}
\concept{api}
