% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Royale.R
\name{set_Royale}
\alias{set_Royale}
\title{Set Clash Royale API key}
\usage{
set_Royale(key)
}
\arguments{
\item{key}{Required. Your Clash Royale API key. See \href{}{}.}
}
\value{
None
}
\description{
Sets your Clash Royale API key for default use.
}
