% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clan.R
\name{cr_get_clan}
\alias{cr_get_clan}
\title{Get a single player}
\usage{
cr_get_clan(clan = "99R2PQVR", key = cr_get_key())
}
\arguments{
\item{clan}{Required. Clan tag. Default: 99R2PQVR}

\item{key}{Required. API key. See <https://developer.clashroyale.com/#/documentation>
Default: cr_get_key}
}
\value{
Returns a player's info as a tibble
}
\description{
Gets full clan details
}
