% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Royale.R
\name{cr_set_key}
\alias{cr_set_key}
\title{Set Clash Royale API key}
\usage{
cr_set_key(key)
}
\arguments{
\item{key}{Required. Your Clash Royale API key. See <https://developer.clashroyale.com/#/getting-started>.}
}
\value{
None
}
\description{
Sets your Clash Royale API key for default use.
}
\examples{
# TODO
}
