% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/royale-package.R
\docType{package}
\name{royale-package}
\alias{royale}
\alias{royale-package}
\title{royale: Clash Royale API wrapper for R}
\description{
R interface to the official API for Clash Royale \url{https://developer.clashroyale.com/#/}.
}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{christopherkenny@fas.harvard.edu} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
